refactor(team): extract read-only queries from broker.go
Phase 3 fourth slice. Pulls 23 read-only accessor methods + the
queueSnapshot type into broker_queries.go. Each takes b.mu, copies
the relevant slice/struct, and returns — no state mutation.

Side-effecting "ensure" methods (EnsureBridgedMember,
EnsureDirectChannel, PostInboundSurfaceMessage, ExternalQueue —
which mutates b.externalDelivered) stay in broker.go because they
write; they are queries in name only.

Methods moved (read-only, side-effect-free):
  Messages, ChannelMessages, AllMessages, OfficeMembers
  EnabledMembers, DisabledMembers, SurfaceChannels, DMPartner
  HasPendingInterview, HasBlockingRequest, HasRecentlyTaggedAgents
  ChannelTasks, AllTasks, InFlightTasks, UnackedTasks
  RecentHumanMessages, Requests
  Actions, Signals, Decisions, Watchdogs, Scheduler
  QueueSnapshot (+ queueSnapshot type)

  - internal/team/broker.go: -324 LOC (2700 -> 2376)
  - internal/team/broker_queries.go: +345 LOC (new)

Cumulative broker.go reduction across Phase 3:
  3404 -> 2859 (web proxy + sse)         -545
       -> 2800 (auth + publish)          -59
       -> 2700 (wiki lifecycle)          -100
       -> 2376 (queries)                 -324
  Net: -1028 LOC; broker.go shrunk by 30%.

No behavior change. Tests pass.
